move utility classes for instance checking out of __init__.py
